feat: ShapeApp: make main function to be standalone
