+++ Shapes/Shapes/RectangleShape.cs
@@ -1,4 +1,4 @@
-﻿namespace ShapeApplication {
+﻿namespace ShapesApplication {
     public class RectangleShape : Shape {
         public double Length { get; set; }
         public double Width { get; set; }
